Apply security/correctness fixes from full code review
Two-pass review (memory-safety bugs + design root causes). Build-verified
clean from scratch (Release|x64, v143).

Critical/High (untrusted-input memory safety):
- dcraw ppm_thumb/ppm16_thumb: compute thumbnail sizes in size_t. The x64
  pixel cap (65535*65535) allowed w*h*3/4 to overflow 32-bit int, yielding
  a tiny allocation overrun by the copy loop -> attacker-controlled heap
  overflow on a crafted RAW thumbnail.
- dcraw_mod: arm setjmp(failure) in the live dcraw_main entry. merror()/
  derror() longjmp through an otherwise-unset global jmp_buf (the only
  setjmp was in the dead CLI main()) -> UB/crash on malformed RAW.
- Helpers::FindEXIFBlock: validate the declared APP1 segment length fits the
  remaining buffer, closing a ~64KB heap over-read in the CJPEGImage EXIF
  copy reachable from any malicious JPEG.

Medium/Low memory-safety & correctness:
- Helpers::SafeImageByteSize: 64-bit image byte-size helper; route SaveImage
  full-size save and clipboard copy. WICLoader: size math in size_t + INT_MAX
  guard, and check the GetSize HRESULT.
- WorkThread/DirectoryWatcher: _beginthread -> _beginthreadex (caller-owned
  handle, CloseHandle once) so Terminate/Abort wait on the real thread.
- ~CImageLoadThread: Terminate() (join) before freeing decoder caches the
  worker may still touch -> closes a shutdown use-after-free window.
- EXIFReader: add lower-bound guards to the IFD0/EXIF/GPS offset checks
  (x86 pointer wraparound OOB read); compare BOM bytes as unsigned char.
- WebP: check WebPDecodeBGRAInto return (uninitialized-heap display);
  nothrow animation-cache allocation.
- Fix '*p++ = f(*p)' unsequenced read/modify UB (ReaderTGA, PSDWrapper,
  6x ImageLoadThread alpha-premultiply loops).
- Clipboard paste: validate untrusted CF_DIB header fields against
  GlobalSize before locating the pixel bits (OOB read).
- ParameterDB v1->v2 conversion: use the bytes actually read, not file size.
- PNG pixel buffer new[] (was malloc, freed with delete[] by CJPEGImage);
  TGA pixel buffer + BMP header zero-init (uninitialized-heap on truncation).
- new[]/delete[] mismatches (FileExtensionsRegistry, PrintDlg).
- dcraw jpeg_thumb: reject oversize length, use bytes read; Sinar tag
  allocate +1 and NUL-terminate to bound the strchr/strncmp scans.
- BasicProcessing: size_t casts on full-image allocation size math.

Design improvements:
- ParameterDB: O(1) unordered_map hash index (rebuilt after each mutation)
  replacing the O(n) per-open linear scan; atomic temp-file + MoveFileEx for
  the full-DB rewrite so a crash mid-merge cannot corrupt the DB.
- ImageLoadThread: single EvictOtherDecoderCaches() helper replacing the
  per-format hand-maintained (and inconsistent) cache-eviction lists.

+++ src/JPEGView/BasicProcessing.cpp
@@ -522,7 +522,7 @@ void* CBasicProcessing::ApplyLDC32bpp(CSize fullTargetSize, CPoint fullTargetOff
 		return Apply3ChannelLUT32bpp(clippedTargetSize.cx, clippedTargetSize.cy, pDIBPixels, pLUT);
 	}
 
-	uint32* pTarget = new(std::nothrow) uint32[clippedTargetSize.cx * clippedTargetSize.cy];
+	uint32* pTarget = new(std::nothrow) uint32[(size_t)clippedTargetSize.cx * clippedTargetSize.cy];
 	if (pTarget == NULL) return NULL;
 	CProcessingThreadPool& threadPool = CProcessingThreadPool::This();
 	CRequestLDC request(pDIBPixels, clippedTargetSize, pTarget, fullTargetSize, fullTargetOffset,
@@ -1314,7 +1314,7 @@ void* CBasicProcessing::RotateHQ(CPoint targetOffset, CSize targetSize, double d
 		return NULL;
 	}
 
-	uint8* pTargetPixels = new(std::nothrow) uint8[targetSize.cx * 4 * targetSize.cy];
+	uint8* pTargetPixels = new(std::nothrow) uint8[(size_t)targetSize.cx * 4 * targetSize.cy];
 	if (pTargetPixels == NULL) return NULL;
 
 	CProcessingThreadPool& threadPool = CProcessingThreadPool::This();
@@ -1465,7 +1465,7 @@ void* CBasicProcessing::TrapezoidHQ(CPoint targetOffset, CSize targetSize, const
 		return NULL;
 	}
 
-	uint8* pTargetPixels = new(std::nothrow) uint8[targetSize.cx * 4 * targetSize.cy];
+	uint8* pTargetPixels = new(std::nothrow) uint8[(size_t)targetSize.cx * 4 * targetSize.cy];
 	if (pTargetPixels == NULL) return NULL;
 
 	CProcessingThreadPool& threadPool = CProcessingThreadPool::This();
@@ -1498,7 +1498,7 @@ static uint8* ApplyFilter(int nSourceWidth, int nTargetWidth, int nHeight,
 						  int nFilterOffset,
 						  const uint8* pSource) {
 
-	uint8* pTarget = new(std::nothrow) uint8[nTargetWidth*4*nHeight];
+	uint8* pTarget = new(std::nothrow) uint8[(size_t)nTargetWidth*4*nHeight];
 	if (pTarget == NULL) return NULL;
 
 	// width of new image is (after rotation) : nHeight
